build: Teach depsgen glob how to get files from filelist
This adds several more parameters:

--mode which defaults to "args", which does what it was doing so far -
getting a list of files from the command line. The other mode is
"filelist".

--filelist which tells depsgen which filelist it should parse. Only
valid in "filelist" mode.

--map-to which tells to which directories the items in the filelist
should be mapped. It can be passed several times if you have several
directories with the same contents.

This is a transitional phase, where both modes are supported. But the
goal is to remove the "args" mode and rely only on filelists. But that
can happen only when all invocations of depsgen glob are changed to
use filelists.