rkt: send headers on redirect
by default golang's http.Client doesn't copy the original request headers on a
http redirect.
This currently breaks ETag handling and authentication (additionally the
default user agent is sent on redirect instead of the rkt one)

This patch defines a checkRedirect function to populate the headers also on
redirect.

The ETag header is sent for every request.

The authentication headers are sent only if there's a related configuration for
the current request's host. For example if authentication is configured for
example.com but not for anotherdomain.com and a request to example.com
redirects to anotherdomain.com, no authentication headers will be provided to
it.