Rework SELinux implementation
We were previously transitioning into the confined SELinux context at the
point we launched stage1, but this works badly - we still need to access a
number of host code at this point. systemd-nspawn supports changing the
context of the app it launches, so switch to using that instead. This is
implemented by using an environment variable to pass the desired context
from the stage0 into the stage1 init app.