build: Prepare for different ldflags syntax in newer go
Go 1.5 deprecated the "-X variable value" format in favor of " -X
variable=value" one. Unfortunately Go 1.4 does not accept the latter
format.

This commit introduces a shell function RKT_XF which outputs the old
format for Go 1.4 or older and a new one for Go 1.5 or newer. My
initial idea was to specify a separator variable which would hold a
space for old Go and = for new Go, so it would be used like "-X
variable${X_SEPARATOR}value". I dropped the idea, because of quoting
issues - we want to be safe against spaces in a value part. That's
why, for old Go, we have "-X variable 'value'" and for new Go - "-X
'variable=value'". Using "-X variable='value'" ended up with version
string enclosed with literal single quotes.

When the support for Go 1.4 is dropped, the code can be then
simplified.