metadata svc: fix typo in the HTTP-POST options
When requesting the metadata service to verify a signature from another pod,
the uuid of the pod needs to be passed as an HTTP-POST option. According to
the spec, the name of the option is "uuid" and not "uid".

Note: the same typo was done in the ACE validator: https://github.com/appc/spec/pull/485