*: change appName to name for images
When storing and printing information about ACIs, we were using the
rather incorrect `appName` instead of the more appropriate `name`. This
changes that in both the `rkt image list` output and in the underlying
store, as well as other internal datastructures.

Since this is one of the columns in the database we use to index ACI
information, it also involves a database schema migration. ql does not
provide primitives to perform this trivially (like copy or rename), so
we achieve this by:

1) creating a new temporary table
2) copying all rows from the original aciinfo to the temporary table
3) dropping the original aciinfo
4) creating a new aciinfo table with our new schema
5) copying all rows from the temporary table to the new aciinfo
6) dropping the temporary table

This allows us to preserve the column ordering and minimise other
changes.

Hat tip to @sgotti for improving the original procedure here.