build: Add a stamp parameter to build_go_bin.mk
This stamp will later group all tasks related to building the go
binary, that is - building the binary, generating filelists and
generating dependencies.

For now, it only has the go binary as a prerequisite, but that will
change later when filelist and deps generation is splitted out to
separate rules.