build: Cleanups in find-so-deps.mk
Rename a variable. Later, we will copy libraries to another directory
named "libs", so we will be able to generate a filelist. To avoid a
confusion with a future variable holding a name of a temporary
libraries directory, we rename the variable holding paths for the
LD_LIBRARY_PATH environment variable.

Also, add a toplevel stamp, which for now depends on another stamp, so
it looks redundant. But later it will depend on more stuff, like
generated clean files.

This is a preparatory commit for adding filelist generation.