build: Cleanups in kernel.mk
The bzImage file might be created somewhere during the make run. To
ensure that the build has finished completely we make the stamp
creation to be the last step in the rule. We will want all the created
files to be there, so we can put them into filelist.

This is a preparatory commit for adding filelist generation.