build: Cleanups in usr_from_src.mk
This commit provides a toplevel stamp that does not do too much for
now - only ensures that ACI rootfs is prepared. But later it will
depend on more stuff like clean.mk and dep.mk to be generated.

This also splits the rule doing the configure, make and make install
into two rules - one running "./configure" and "make", and the other
running "make install-strip". This split is also dictated by adding
filelist generation in future commits - one filelist will be generated
for the build directory and one for the installation directory.

This is a preparatory commit for adding filelist generation.