build: Add filelist generation to find-so-deps.mk
For now, nothing depends on the generated filelists, so the generation
won't happen. Later, deps and clean file generation will depend on
them.

Libraries are copied into two directories - to a temporary libs dir
and to ACI rootfs. The temporary libs dir contains only copied
libraries (as the directory in ACI rootfs may already contain some
libraries before the copying started), so we can generate a filelist
based on this directory. Later the filelist can be used to generate
clean file for deleting copied libraries in both directories.