build: Add filelist generation to usr_from_coreos.mk
For now, nothing depends on the generated filelists, so the generation
won't happen. Later, deps and clean file generation will depend on
them.

The contents of squashfs file are put into some temporary directory
and then they are copied to ACI rootfs. That way we get an exact list
of files taken from squashfs. The filelist is generated basing on the
contents of the temporary directory.