stage0: added ability to resume interrupted downloads
Downloads for a given url are now saved to a file named after a sha512
hash of the url. If the file already exists for a requested download, a
check will be made to see if the server supports HTTP range requests. If
such requests are supported then only the missing part of the file will
be downloaded, otherwise the entire file will be downloaded as before.