rkt: Make run-prepared rerunnable when mds is missing
When a user has no metadata service running and runs rkt run-prepared
without specifying mds-register=false, the pod is left in the run state
and the user needs to reprepare the pod again to rerun.

To avoid this scenario, this commit does an early check to see if the
metadata service is available before the pod is moved to the run state.

fixes https://github.com/coreos/rkt/issues/1434