lkvm: make sure $HOME is defined before running lkvm
kvmtool currently does not work when $HOME is not defined. When started
by a systemd unit file (or by systemd-run), $HOME is typically
undefined, unless "User=" (or resp. "--uid=") is given.

Instead of asking users to work around the issue with "User=" or
"--uid=", just define that variable in rkt before calling lkvm.

This could be removed when this gets fixed in lkvm upstream.

Issue described in https://github.com/coreos/rkt/issues/1393