rkt: Provide more readable output to rkt image list
The output of rkt image list was not very pleasant due to the very long
hash and data format. This change shortens the hash to show the hash
identifier plus the first 12 alphanumeric digits after the '-' char by
default. The date format is changed to use a human readable format such
as "5 days ago".

If one care to have the older output format, she can use the newly
introduced --full flag. Similar to the --full flag in rkt list, this
provides a full hash and date output string.

A follow up change needs to be made so that the commands that take the
image hash as an argument can use the short form. Currently the short
version of the hash is of little use other than viewing.

Closes https://github.com/coreos/rkt/issues/1438
Closes https://github.com/coreos/rkt/issues/1439