semaphore: Update install-deps
We don't need to set gcc via alternatives to version 4.8 - it is
already fixed. Also, instead of installing go and gofmt with apt-get,
just use what semaphore provides with change-go-version script.

Maybe makes the builds a bit faster.

Using change-go-version does not seem to work in install-deps, so try
calling it in run-build.sh script.