rkt: Get rid of unused named return variable
We usually don't use this style. Also, in one line we did not return
any value, so a default integer value was returned instead (as the
return variable was not set to anything, so it defaulted to 0). This
meant returning 0 even if we could not get ACI infos.

Also removed the same from pod list command.