rkt: Remove tabOut global variable
This replaces a tabOut global variable with a function that sets up a
new tab writer consistently, but allows caller to customize the
backing io.Writer.

In later commits, the `rkt image list` and `rkt list` commands will
change the backing writer to the bytes buffer to ensure that the list
is written after all the errors.