build: Fix a race
There was a race between installing lkvm or kernel and removing the
whole rootfs.

This is not an entirely proper fix, such will land later. But should
be good enough.

Both lkvm and kernel were installed around the same time as ACI rootfs
was prepared. But rootfs was removed as a part of the preparation, so
if the installation came first, then both lkvm and kernel (or just one
of them) was removed and not installed again.

Put kernel and lkvm binary into STAGE1_INSTALL_FILES, which will do
the installation after the rootfs preparation.