build: Fix another race
os-release file was an odd-man-out - it was not copied, but created
directly in ACI rootfs. Also, it did not depend on ACI rootfs initial
contents to be created, so it was possible for rootfs to be cleaned up
after os-release was created there.

Fix it by adding simple os-release file to repo and putting it into
ACI rootfs by usual STAGE1_INSTALL_FILES variable.