functional tests: Avoid rebuilding ACE images on every run
There is a hack which sets timestamps of some files to epoch to get
consistent hash of some kind. These timestamps were making the
dependencies of this rule to be older than target, thus resulting in
rebuilding the target.

The purpose of this hack is not clear and for sure it was not working,
because we were not setting the timestamp of the other directory
(/opt/acvalidator), so the hash would not be consistent anyway.

Just remove it.