api: change version from v1 to v1_alpha
The API that landed in a609b37fd28679fe76c561d34ea14632aeaa1741 is
highly experimental and subject to change at any time, but we've already
had users wondering if they should be consuming it. Add a big scary
warning to make it very clear that they should not yet.