build: Fixes for older autotools
Centos 6.7 has buggy autoconf/m4/whatever - it is getting confused
with empty conditional bodies preceded with m4 comments, which results
in generating shell code with empty conditional bodies, which in turn
is an invalid code.

Just put some colons (which serve as an explicit noop in shell) to be
safe.

Also, lower autoconf required version.