tests: Create ACI with different hashes
The "Multiple apps" was using the exact same image to run the test but
expecting different hashes to be returned. Because the same hash was
being returned and a `rkt image rm` command was being issued for each
hash, the second attempt to remove the image was failing.

This change applies a simple name patch to each image to force a unique
hash.