stage1: Rename the manifests
Add the ".manifest" extension - it will be needed for generating
shallow filelists.

This is rather a workaround - a proper fix would be to add a flag to
"depsgen glob" specifying a shallow or deep mode instead of deciding
about it based on whether there is a suffix (empty suffix - deep mode,
non-empty suffix - shallow mode).