build: Split off common CoreOS variables into separate file
The coreos-common.mk contains the rules for downloading and unpacking
the image to get the squashfs file. This stuff is split out, because
it will need to be executed only once, even if building several
flavors using it (coreos and kvm ones). To make sure that we do not
end up with the two same rules for the pxe.img (which would earn us a
warning from make), coreos-common.mk has C-like inclusion guards.