rkt: provide image version and hash in list output
The `rkt list` command currently shows an inadequate amount of
information to determine which image an app is associated with. For
example, one could have fetched multiple versions of an image and it's
possible for images to have the same name but different hashes. However,
these details are not shown in the output.

This change adds the version information to the default `rkt list`
output and adds the 12-char ID of the image to the output when the
'--full' flag is used.

It also changes the 'ACI' field header to 'IMAGE NAME'.

Closes https://github.com/coreos/rkt/issues/1492