stage0: make options output nicer
The options section in the help output was not formatted with spaces
between the name/command and its description. This change adds a
function, mostly copied from the pflags source, which adds a tab instead
of the ": " before the description. The tabwriter takes care of the
rest.

It would be nice to have a change upstream in pflags and/or cobra which
would make this less hacky.

closes #1113