api: change the filter.
Add 'ids' field in the PodFilter.

Change 'image_names' to 'image_ids' field in PodFilter, as ListPods()
can return pods that only have image ids in the manifest.

Change the filtering rules for annotations: In order to be returned,
the pods/images only need to have one of the listed annotations,
instead of all of them.

Change the 'api_version' in Info from integer to string.

Also make the description of the filters more consistent.