directory parameters: allow relative paths
The parameters --dir=, --system-config= and --local-config= used to only
accept absolute paths. This patch changes the type of the parameters
from string to absDir. When the flag is parsed, the potentially relative
path is converted to an absolute path automatically.

The parameters still have the same defaults. It is no longer allowed to
pass an empty string (e.g. --dir="" is not allowed).

Fixes https://github.com/coreos/rkt/issues/1605