pkg/tar: rename extractTar to ExtractTarInsecure
It would be useful in acbuild to be able to untar ACIs without
chrooting, so this commit makes the function that does the extracting of
the tar file exposed. With this change the tar package exposes an
ExtractTar function for extracting an ACI inside of a chroot, and an
ExtractTarInsecure function for extracting an ACI without any chroot
happening.

Additionally the ExtractTarInsecure function's logic related to the
uidshift has been factored out into a function that gets passed in, so
that clients using it can have better control over it.