stage1: rename diagexec -> appexec
When systemd encounters errors executing apps via diagexec, the log
messages contain spurious mention of this "diagexec" thing that
end-users don't have any familiarity about.

The utility of diagexec has expanded beyond just wrapping the execution
of the app with a diagnostic on the failure case, and a name like
"appexec" is less ambiguous to the end-user.