stage0/run: use store.GetImageManifestJSON() instead
This way we don't need to re-marshal the manifest into a []byte, just
trust the manifest coming from the store.  If the manifests need
validating we should do that before putting them in the store.  This is
in the hot path of starting pods since we write manifests for the app
and stage1 acis.