coreos: Small fixes in cache.sh
Variables in functions should be marked as local. Also looks like the
form "local var=foo" should not be used, because of a strange order of
processing this line - first var is marked as global variable and a
value is assigned to it, then the variable is made local.

Also, make it run with "set -e", so it will bail out if downloading or
verification failed.