stage0: break-up --insecure-skip-verify
Until now security features were either all on or all off. This commit
provides a means to turn specific security features on or off. By
default security features are all active.

This means that we've replaced the --insecure-skip-verify flag with one
called --insecure-options. This flag accepts a comma separated list of
features to turn off. The options that are currently accepted are
'none', 'image', 'tls', and 'all'. 'none' is the default and 'all' will
give one the previous behaviour from --insecure-skip-verify.

A BitFlags type was created to accept this list and provides a means to
check if a feature is active or not. For convenience, wrapper methods
are provided to query a feature's state; SkipImageCheck(),
SkipTlsCheck(), SkipFingerprintCheck(), SkipAllSecurityChecks(), and
SkipAnySecurityChecks().

Skipping the fingerprint review in `rkt trust` has also been moved to a
flag specific to the `trust` subcommand instead of being included in the
global --insecure-options flag.