stage0: don't change the group ownership of rootfs
Setting the group of the rootfs to 'rkt' causes problems with user
namespaces when the group is not mapped in the user namespace. It
prevents the container from doing a 'mkdir' or a 'lstat' on /proc. See
https://github.com/systemd/systemd/issues/1585

Setting the group ownership of the rootfs to 'rkt' was done in
https://github.com/coreos/rkt/pull/1452 so that the command 'rkt status'
could work as non-root. However, if the rootfs is r-x for others, setting
the group should not be necessary. r-x for others was removed by
https://github.com/coreos/rkt/pull/1607 but I am adding it back.

This patch reverts a part of https://github.com/coreos/rkt/issues/1602,
therefore fixing the regression with user namespaces. 'rkt status' as
non-root still works.

Fixes https://github.com/coreos/rkt/issues/1602