stage0: Remove use of ImageManifest in rkt list
Since we are now able to remove an image without removing the pod, use
of the ImageManifest in the rkt list creates a dependency on an image
that may not be around, causing rkt list to fail.

This change removes that dependency and instead relies on the
PodManifest which will be available and should supply the needed
information.