rkt: remove `--inject-volume` from run/prepare
The syntax of `--inject-volume` has been somewhat [contentious][1] and
it seems pretty clear we do not want the flag to remain as-is.
Furthermore, there's actually no immediate need for it since the same
functionality can be achieved through a combination of the `--volume`
and `--mount` flags (which landed in the [same PR][2] as `--inject-volume`).
Hence, remove it for now, guide people towards the alternative, and we
can put further thought into re-adding something similar in future
(which would essentially just be a convenience).

[1]: https://github.com/coreos/rkt/issues/1656
[2]: https://github.com/coreos/rkt/pull/1582