build: Fix ACI Go tools not being rebuilt on the source file changes
stage1 ACI image had an indirect dependency on a tool (init or gc)
inside the ACI rootfs directory, but it didn't have any dependency on
the stamp which, besides making sure that the binary was built, is
making sure that the deps file are generated. Those deps file are
adding dependencies to built binary on its source files.

Make the binary in the ACI rootfs directory to depend on the stamp
too, so the missing files are generated.