rkt: get image name from appinfo ImageManifest
Currently we are relying on an optional field in the PodManifest to get
the images name and version when running `rkt list` because, since one
has been able to remove referenced images, there was no guarantee that
the ImageManifest in the cas would be available.

However, little did we know, there is a copy of the ImageManifest of the
pod in the pod's appinfo directory. This commit makes use of that.

fixes #1659