stage1: pass SYSTEMD_NSPAWN_CONTAINER_SERVICE env var
With https://github.com/systemd/systemd/pull/1817, systemd-nspawn can
receive the name of the container service that's calling it.

This commit passes "rkt", which allows the container to be registered to
machined with the "rkt" name.

Note that this will start working when we update stage1's systemd to
v228, but doing it now doesn't hurt and we'll be ready.