functional tests: fix TestFetch
We now generate an implicit empty volume if a mount point is specified
in the image but no volume is passed to rkt. Some test cases in
TestFetch were relying on the fact that rkt doesn't actually run these
images.

Fix the test by adding an `--exec /dev/null` to make the images not run.