CHANGELOG.md: add an entry for unreleased changes
Updating the changelog at each release takes time because we need to go
over all the changes since the previous release. I would like to add a
"vUNRELEASED" entry to update as long as we implement something worthy
of writing in the changelog. Then, we would just rename the
"vUNRELEASED" entry during the release process.