functional tests: Fix racy ACE validator test
The "main" and "sidekick" stages of ACE validator run in
parallel. Most of the time the "main" stage is first with printing its
OK message, but sometimes the "sidekick" wins the race. In the latter
case the test failed.

Now the test expects "prestart" to be first, followed up by "main" and
"sidekick" in any order and then "poststop" to be last.