build: Add --with-stage1-flavors-version-override
This flag overrides the version used in built stage1 image
manifests. Also, when using --with-stage1-default-flavor, the new flag
changes the default stage1 version to the overridden one, so rkt will
search in the store for an image with the standard coreos name and
with an overridden version.

This flag is useful for distros serving updated revisions of rkt, like
for example in Debian, they might want to put a version 0.12.0-2.