docs: Correct a wishful thinking in hacking
rkt usually will not be a static binary. We tried once and the result
has problems with networking. It is probably like this because
statically linking glibc into a binary makes the glibc not to use its
plugin based name service switch fully, which is responsible, among
others, for resolving domain names.