stage0: brute-force GC on GC failure
When a pod directory is in an unexpected state, it may happen that the
standard garbage collection fails. When this happens mounts can be left
around, requiring manual removal.

This change forces a GC of the pod mount points when the pod GC
otherwise fails. To do this, the /proc/self/mountinfo file is parsed,
the pod mount points are extracted, ordered, then unmounted.

fixes #1572