rkt: add image size to rkt image list
We store the size of the ACI in the database when we write it to the
CAS. We also store the size of the tree store when we render the ACI.
We also populate sizes of existing images after we do the migration to
the new db schema.

When we list images, we print the sum of these two sizes as the size of
the image. This means that if the image is fetched but never ran, it
won't be rendered to the tree store so it will take less space.

If the image is removed (`rkt image rm`), we also remove the ACIInfo
table associated with it, this means we don't have information about the
tree store size. It will be removed on the next garbage collection if
it's not referenced by any non-GCd pod.