docs: rework README + add docs for nomad/k8s
This breaks apart much of the README into a new "basics" document. The
goal is to make the README a little more approachable and do better at
highlighting the different options for people to reach out.

Also adds initial documents describing integration with Nomad and
Kubernetes and links to them appropriately.