api_service: Minor refactoring on what to return for different pod state.
If a pod is in Embryo, Preparing, AbortedPrepare state,
only id and state will be returned.

If a pod is in other states, the pod manifest and
apps will be returned when 'detailed' is true in the request.

Valid pid and networks are only returned when a pod is in Running.