MAINTAINERS: remove eyakubovich, add dgonyeo
Eugene is sadly no longer with CoreOS, so remove his handle from
the list of maintainers. He will be gladly welcomed back if he
decides to continue contributing to the project.

In happier news, Derek has joined us as an active and effective
contributor, so add him to the maintainers.