functional tests: Drop some more dead code
The stop channel is redundant. Instead, just listen to the closing of
the message channel.

Also, it was possible to build the ACI auth server as an executable
binary. That was from the times, when I initially wrote it as a
standalone tool. We don't use it anymore - the project also provides a
library, which is better suits our tests better.