api_service: Use image manifest in pod directory for List/Inspect pod.
This commit changes the way how List/Inspect pod fetches image
manifest for apps.

Previously, the manifest is read from store, which will be empty
if the image is removed. Now the manifest is read from the image
manifest inside the pod directory, so it won't be affected by
image removal.

However, some information like importtime, image size are not
available for List/Inspect pod after this change.