rkt: handle all output through new log package
This change creates a new type, Logger, that wraps go's own Logger type
to provide consistent output across go commands. Along with the methods
from go's log package, there are also methods that accept our structured
error type introduced in the previous commit. These commands append an
'E' to the end of the familiar methods. For example, we now have
'PrintE', 'PanicE', & 'FatalE'.

With this, we have one, central location to modify the way output is displayed in
rkt.

The output generated when printing errors is:

```
cat-manifest: problem retrieving pod: no matches found for "asf"
```

or when using the `--debug` flag we'll get the entire error stack.

```
cat-manifest: problem retrieving pod
  └─unable to resolve UUID
    └─no matches found for "asf"
```

From this example we can see that when no `--debug` flag is used, you
get the innermost error as the last output on the line. All other errors
are hidden. This text between the 'command name' is the programmer
defined message that accompanies the errors sent to output. This can be
left empty.

When using the `--debug` flag we only get the 'command name' and
accompanying message on the first line. Under that we get the unwound
error messages, with the innermost one at the bottom.

Furthermore, all output to stderr has a prefix that is either
programatically set using the command or subsommand name, or one set in
the package.

refs: #1550
fixes: #1754