stage1, pkg: add acl package
This package is a wrapper over libacl that, instead of being linked to
it, it dlopens it so we can check if it's present on the system.

It's not meant to be complete, we just implement the functions needed by
rkt at the moment.

It is based on https://github.com/naegelejd/go-acl