functional tests: Add a stub stage1 image for testing
This implementation is actually so minimal, that it does not even
bother with running stage2.

We can use it regardless of how the project was configured. This
stage1 image could also be used in other tests that do not involve
checking things happening inside stage1 or stage2.

This will be used in stage1 loading test.