functional tests: Change the path to rkt binary
The stage1 loading test will use stub stage1 image for its checks. One
of them is testing whether rkt falls back to fetching stage1 image
from the same directory the rkt binary is in. To do this, we need to
put the stub stage1 image in the same directory as rkt. To avoid a
confusion from packagers and users, we do it backwards - we put a
symlink to the rkt binary in the directory where stub stage1 image is.

In the same directory we also put symlinks to all the flavors of the
stage1 image, so the tests should work as they did so far.