trust: only fetch pubkeys when no key for matching prefix
This commit changes rkt's behavior such that when fetching an image
through meta discovery, rkt will only download and trust the gpg public
key for the image if rkt doesn't have any local keys with a matching
prefix. If the key for an image has changed, the user must manually
trust the new key with `rkt trust --prefix`.

In addition, `--trust-keys-from-https` now defaults to false, so the
user is prompted to approve new keys (unless rkt isn't running in a tty,
in which case an error is printed).