pkg: New group package
The group package looks up given group name. It is moved to a separate
package, so we don't have to include the github.com/coreos/rkt/common
package it in github.com/coreos/rkt/tests/testutils. Importing it
there ended up in errors when building the inspect binary, which is
importing testutils. common package was importing indirectly some
other package, which requires cgo, but cgo was disabled when building
inspect.