stage1: fix socket activation port matching
We were using the pod's port to set up the socket unit file inside the
pod. This means it only worked if the port in the image manifest and the
port in the pod manifest (set by rkt at runtime) were the same.

The reason is that systemd inside the host matches based on the socket
port number, and since the socket is created on the host, it will have
the host port number. If this number is different from the one in the
image manifest, systemd will not assign the socket to the correct app in
the pod because it will expect a socket with the image manifest port
number.

The solution is using the host port number for the socket unit file.