stage1/fly: fix error with unset exec statement
If the fly stage1 was passed an image that had an empty exec, it could
panic. This adds the same check that we use for the non-fly stage1, and
refactors the code slightly to share it between them.

Also renames a couple of old vestigial variables: appmanifest rather
than imagemanifest, etc.