functional tests: Workaround a listening on localhost failure on ubuntu 16.04
This commit fixes one problem, and works around another one.

The first problem was assuming that localhost = 127.0.0.1. While this
is a quite safe assumption, one can always do weird stuff in their
/etc/hosts (like no ipv4 for localhost). Instead we look up the
address of localhost and try to listen to it.

The problem we are working around is a some kind of race somewhere -
the old server closes the listener and then the new server creates a
new one which listens on the same address and port. Somehow, something
in the network stack acknowledges the closing of the listener with a
delay, which causes new server to fail to listen on that address and
port, because it is still bound. This has happened only on Ubuntu 1604
so far.

There wouldn't be any such problems if we had discovery on a custom
port instead, because we wouldn't have to listen on fixed ports (like
80 or 443).