build: Fix a warning printed when checking go version
For certain go versions (1.5, 1.6rc2), the GO_MICRO variable was
empty, that is - it wasn't a valid integer expected by shell for its
numeric relation operators (-lt, -eq and so on). We are using these
operators for checking if go is without CVE-2015-8618. The `if`
statement after the expansion looked like follows:

`if test "1" -eq "1" -a "6" -eq "5" -a "" -lt "3"`.

The warning was printed even if the `"" -lt "3"` should not be
evaluated, because `"6" -eq "5"` is false. Looks like bourne shell is
doing only a cursory short-circuiting, if any.