stage1: propagate app exit code to rkt exit code
If the version of the systemd running in stage1 is >=v227, the command
"systemctl exit STATUS" will set the exit status of the container. See
https://lists.freedesktop.org/archives/systemd-devel/2015-October/034509.html:

 * The "systemctl exit" command now optionally takes an
   additional parameter that sets the exit code to return from
   the systemd manager when exiting. This is only relevant when
   running the systemd user instance, or when running the
   system instance in a container.

If not, we will continue with the previous logic, so we can merge this now
and benefit from it when we actually update to v229.