prepare-app: create /dev/log for stage2
rkt starts systemd-journald in the pod and creates the symlink in the
stage1 rootfs:
/dev/log -> /run/systemd/journal/dev-log (socket file)

But unfortunately it isn't directly available for the app because the
app is chrooted.

systemd-journald listens on the following sockets and other paths:
- /dev/kmsg,
- /dev/log,
- /run/systemd/journal/dev-log,
- /run/systemd/journal/socket,
- /run/systemd/journal/stdout

This patch prepares /dev/log in stage2 as a symlink to
/run/systemd/journal/dev-log and bind mounts the directory
/run/systemd/journal from stage1 to stage2, so both /dev/log and
journald's native logging will work for the application

This will help haproxy and nginx.

See https://groups.google.com/forum/#!topic/rkt-dev/P6HEylgrv8U