rkt/image: reset session on bad HTTP status
When figuring out if we can resume a session, if we get a bad HTTP
status, we were just returning without error and weren't resetting the
session. This means we will download the full file *on top* of the already
present file, which of course leads to a corrupted file.

Reset the session in that case.