build: Fix dependency generation for go binaries
The way depsgen works is that it recursively gets all the dependencies
of a given module (say, github.com/coreos/rkt/rkt), filters the
dependencies to eliminate packages outside the repository (basically
packages from Go standard library), and then gets the files of all
those filtered deps and the module. Filtering was done by simply
checking if an import path of a dependency had a specific prefix,
namely a name of the repository (in our case the repo name is
github.com/coreos/rkt). This worked fine until we stopped rewriting
our imports. It is because rewritten imports always had the prefix we
wanted (github.com/coreos/rkt/Godeps/_workspace/etc/etc) and now they
do not. In effect, we were getting the files of all the dependencies
that are a part of the repository (like github.com/coreos/rkt/stage0).

So now, instead of checking ImportPath of a dependency, we check its
Dir and check if it is a direct or indirect subdirectory of the repo.